+++ changelog.md
@@ -1,4 +1,5 @@
 # Changelog
 
-## 01-03-2022
+## 02-02-2022
+- Update dummy events for demo
 - Fix is-today color on first week of month view